+++ test_settings.py
@@ -3,7 +3,7 @@ from foundry.settings import *
 
 DATABASES = {
     'default': {
-        'ENGINE': 'django.db.backends.sqlite3',
+        'ENGINE': 'django.contrib.gis.db.backends.spatialite',
         'NAME': 'test_foundry.db',
         'USER': '',
         'PASSWORD': '',