+++ CHANGELOG.rst
@@ -1,6 +1,21 @@
 Changelog
 =========
 
+1.0
+---
+#. Patch django.contrib.sites.models.Site.__unicode__ so it returns name and not domain. The UI gets confusing since we have up to three sites comprising one logical mobi site.
+#. Listings now have automatic RSS feeds.
+#. Comment form now fires up correct virtual keyboard for a smart phone.
+#. Logged in members can now flag offensive comments. After three flags a moderator is notified.
+#. Some IP addresses can now be allowed to bypass the age gateway / private site.
+#. Listing gets an optional RSS feed.
+#. Simplified paginator. No more breadcrumbs.
+#. Show less metadata in mobi listings.
+#. Ditch addthis sharing widget. It is too slow.
+#. Simplified commenting and chatroom. Removed some navigation links.
+#. Some user agents can now be allowed to bypass the age gateway. This allows bots to crawl the site.
+#. Up required jmbo to 1.0.
+
 0.7.2
 -----
 #. Hotfix. Apps with empty URL patterns cause infinite recursion when adding a page.